Move createDatabase to its own file, databaseUtils.ts.
Also use createActionResultFromMarkdownDisplay from typeagent.
